+++ Project_Dir/src/main/java/com/telerikacademy/carservice/models/Customer.java
@@ -11,7 +11,7 @@ public class Customer {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "customerId")
-    private int customerId;
+    private Long customerId;
 
     @Column(name = "email")
     @Email
@@ -39,11 +39,11 @@ public class Customer {
         this.name = name;
     }
 
-    public int getCustomerId() {
+    public Long getCustomerId() {
         return customerId;
     }
 
-    public void setCustomerId(int customerId) {
+    public void setCustomerId(Long customerId) {
         this.customerId = customerId;
     }
